reorganized stacks
